Fix stringp error caused by not nil *print-length*.
Uruk returns processed documents as list of strings.  *print-length*
clojure repl setting will cat the rest of any sequence after this
element number.  As notification to this event to user clojure repl
replace those elements with three dots.  When handler reads this
sequence, it produce list of strings and a symbol named \.\.\.  Then
uruk handler tries to operate on this last element as a string and
fail with type error.

As a solution to this we overwrite *print-length* and *print-level*
settings to nil explicitly.  To regular cider repl from this change we
create new nrepl session for uruk commands only.  This way we preserve
user settings in the user repl and ignore them in our special session.